bIndexDb is not part of options
git-svn-id: https://svn.code.sf.net/p/comet-ms/code@1124 299d2d02-8f3b-485e-a07e-71e8e5e16577

+++ trunk/comet-ms/CometSearch/CometIndexDb.cpp
@@ -213,7 +213,7 @@ bool CometIndexDb::CreateIndex(void)
       // add composition
       for (int i=0; i<26; i++)
          (*it).iAAComposition[i]=0;
-      for (int i=0; i<strlen((*it).szPeptide); i++)
+      for (int i=0; i<(int)strlen((*it).szPeptide); i++)
          (*it).iAAComposition[(*it).szPeptide[i] - 65] += 1;
       fwrite(&(*it), sizeof(struct DBIndex), 1, fptr);
    }
